Done 18/15 :^)

+++ adcDriver.hpp
@@ -1,8 +1,7 @@
 #ifndef ADC_DRIVER_H_
 #define ADC_DRIVER_H_
 
-#include <stdio.h>
-#include "io.hpp"
+#include "LPC17xx.h"
 
 
 class ADCDriver
@@ -16,10 +15,10 @@ public:
         ADC_PIN_1_31,       // AD0.5
       
         /* These ADC channels are compromised on the SJ-One, hence you do not need to support them
-        ADC_PIN_0_23 = 0,   // AD0.0
-        ADC_PIN_0_24,       // AD0.1
-        ADC_PIN_0_3,        // AD0.6
-        ADC_PIN_0_2         // AD0.7
+		   ADC_PIN_0_23 = 0,   // AD0.0
+		   ADC_PIN_0_24,       // AD0.1
+		   ADC_PIN_0_3,        // AD0.6
+		   ADC_PIN_0_2         // AD0.7
         */
     };
 
@@ -27,30 +26,30 @@ public:
     ADCDriver();
     
     /**
-    * 1) Powers up ADC peripheral
-    * 2) Set peripheral clock
-    * 2) Enable ADC
-    * 3) Select ADC channels
-    * 4) Enable burst mode
-    */
+	 * 1) Powers up ADC peripheral
+	 * 2) Set peripheral clock
+	 * 2) Enable ADC
+	 * 3) Select ADC channels
+	 * 4) Enable burst mode
+	 */
     void adcInitBurstMode();
 
     /**
-    * 1) Selects ADC functionality of any of the ADC pins that are ADC capable
-    * 
-    * @param adc_pin_arg is the ADC_PIN enumeration of the desired pin.
-    *
-    * WARNING: For proper operation of the SJOne board, do NOT configure any pins
-    *           as ADC except for 0.26, 1.31, 1.30
-    */
+	 * 1) Selects ADC functionality of any of the ADC pins that are ADC capable
+	 * 
+	 * @param adc_pin_arg is the ADC_PIN enumeration of the desired pin.
+	 *
+	 * WARNING: For proper operation of the SJOne board, do NOT configure any pins
+	 *           as ADC except for 0.26, 1.31, 1.30
+	 */
     void adcSelectPin(ADC_PIN adc_pin_arg);
     
     /**
-    * 1) Returns the voltage reading of the 12bit register of a given ADC channel
-    *
-    * @param adc_channel_arg is the number (0 through 7) of the desired ADC channel.
-    */
-    uint16_t readADCVoltageByChannel(uint8_t adc_channel_arg);	
+	 * 1) Returns the voltage reading of the 12bit register of a given ADC channel
+	 *
+	 * @param adc_channel_arg is the number (0 through 7) of the desired ADC channel.
+	 */
+    float readADCVoltageByChannel(uint8_t adc_channel_arg);	
 };
 
 #endif
